requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
openai
google-generativeai
pdfkit
jinja2